+++ ichnaea/api/locate/tests/test_internal.py
@@ -1,6 +1,6 @@
 from ichnaea.api.locate.internal import InternalCountrySource
 from ichnaea.api.locate.tests.base import BaseSourceTest
-from ichnaea.region import regions_for_mcc
+from ichnaea.region import GEOCODER
 from ichnaea.tests.factories import (
     CellFactory,
     WifiShardFactory,
@@ -13,7 +13,7 @@ class TestCountrySource(BaseSourceTest):
     api_type = 'country'
 
     def test_from_mcc(self):
-        country = regions_for_mcc(235)[0]
+        country = GEOCODER.regions_for_mcc(235)[0]
         cell = CellFactory.build(mcc=235)
         query = self.model_query(cells=[cell])
         results = self.source.search(query)
@@ -25,7 +25,7 @@ class TestCountrySource(BaseSourceTest):
         ])
 
     def test_ambiguous_mcc(self):
-        countries = regions_for_mcc(234)
+        countries = GEOCODER.regions_for_mcc(234)
         cell = CellFactory.build(mcc=234)
         query = self.model_query(cells=[cell])
         results = self.source.search(query)